// src/app/generated/dcc/model/calibrationCertificateDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MeasurementResultDto } from './measurementResultDto';
import { SignatureDto } from './signatureDto';
import { AdministrativeDataDto } from './administrativeDataDto';
import { ByteDataDto } from './byteDataDto';


export interface CalibrationCertificateDto { 
    id?: string;
    administrativeData?: AdministrativeDataDto;
    measurementResults?: Array<MeasurementResultDto>;
    comments?: Array<string>;
    document?: ByteDataDto;
    signatures?: Array<SignatureDto>;
    schemaVersion?: string;
}