// src/app/generated/dcc/model/conditionDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { DataDto } from './dataDto';
import { RichContentDto } from './richContentDto';


export interface ConditionDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    name?: LanguageSpecificStringsDto;
    description?: RichContentDto;
    status?: string;
    certificateValue?: string;
    certificateId?: string;
    referralId?: string;
    certificateReferral?: LanguageSpecificStringsDto;
    certificateProcedure?: string;
    data?: Array<DataDto>;
}