// src/app/generated/dcc/model/contactDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { LocationDto } from './locationDto';


export interface ContactDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    name?: LanguageSpecificStringsDto;
    phoneNumber?: string;
    location?: LocationDto;
    emailAddress?: string;
}