// src/app/generated/dcc/model/equipmentDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { SoftwareDto } from './softwareDto';
import { ContactDto } from './contactDto';
import { IdentificationDto } from './identificationDto';


export interface EquipmentDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    classId?: string;
    classReference?: string;
    name?: LanguageSpecificStringsDto;
    model?: string;
    manufacturer?: ContactDto;
    identifications?: Array<IdentificationDto>;
    software?: Array<SoftwareDto>;
    certificateValue?: string;
    certificateId?: string;
    certificateReferral?: LanguageSpecificStringsDto;
    certificateReferralId?: string;
    certificateProcedure?: string;
}