// src/app/generated/dcc/model/measurementResultDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { SoftwareDto } from './softwareDto';
import { ConditionDto } from './conditionDto';
import { RichContentDto } from './richContentDto';
import { StatementDto } from './statementDto';
import { EquipmentDto } from './equipmentDto';
import { MethodDto } from './methodDto';
import { ResultDto } from './resultDto';


export interface MeasurementResultDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    name?: LanguageSpecificStringsDto;
    usedSoftware?: Array<SoftwareDto>;
    usedMethods?: Array<MethodDto>;
    description?: RichContentDto;
    results?: Array<ResultDto>;
    influenceConditions?: Array<ConditionDto>;
    equipment?: Array<EquipmentDto>;
    statements?: Array<StatementDto>;
}