// src/app/generated/dcc/model/methodDto.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageSpecificStringsDto } from './languageSpecificStringsDto';
import { RichContentDto } from './richContentDto';


export interface MethodDto { 
    id?: string;
    refIds?: Array<string>;
    refTypes?: Array<string>;
    name?: LanguageSpecificStringsDto;
    norms?: Array<string>;
    description?: RichContentDto;
}